# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: RAG-Based Hindi QA System API
  description: API Gateway for Ministry of Culture conversational chatbot and semantic search
  version: 1.0.0
  contact:
    name: Ministry of Culture, Government of India
    email: arit-culture@gov.in
servers:
  - url: https://culture.gov.in/api/v1
    description: Production API Gateway
  - url: http://localhost:8000/api/v1
    description: Local development API
security:
  - BearerAuth: []
paths:
...
  /search:
    post:
      summary: Unified semantic search
      description: |
        Full-text and semantic search across all 30 Ministry platforms.
        Returns ranked results with AI-generated summaries and multimedia.
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /search/suggest:
    get:
      summary: Search suggestions
      description: Get auto-complete suggestions based on query prefix
      tags:
        - Search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Query prefix
        - name: language
          in: query
          schema:
            type: string
            enum: [hi, en]
          description: Language code
      responses:
        '200':
          description: Suggestions returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /translate:
    post:
      summary: Translate text
      description: |
        Translate text between Hindi and English.
        Supports batch translation for efficiency.
      tags:
        - Translate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
      responses:
        '200':
          description: Translation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /translate/batch:
    post:
      summary: Batch translation
      description: Translate multiple texts in a single request
      tags:
        - Translate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTranslateRequest'
      responses:
        '200':
          description: Batch translation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchTranslateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /admin/analytics/search:
    get:
      summary: Search analytics
      description: Get metrics on search queries, result quality, performance
      tags:
        - Admin
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Analytics data returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAnalytics'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  schemas:
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 1000
        language:
          type: string
          enum: [hi, en]
        page:
          type: integer
          default: 1
          minimum: 1
        page_size:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
        filters:
          type: object
          properties:
            source_sites:
              type: array
              items:
                type: string
            content_type:
              type: string
              enum: [webpage, pdf, document, image]
            date_from:
              type: string
              format: date
            date_to:
              type: string
              format: date
      required:
        - query
        - language
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        multimedia:
          type: array
          items:
            $ref: '#/components/schemas/MultimediaResult'
          description: Images and videos matching the query
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
          description: Relevant events and announcements
        total_results:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        cached:
          type: boolean
...
    SearchResult:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
        snippet:
          type: string
          description: AI-generated summary
        score:
          type: number
        source_site:
          type: string
        language:
          type: string
        content_type:
          type: string
        thumbnail_url:
          type: string
          format: uri
        published_date:
          type: string
          format: date
    MultimediaResult:
      type: object
      properties:
        type:
          type: string
          enum: [image, video]
        url:
          type: string
          format: uri
        alt_text:
          type: string
        source_site:
          type: string
        thumbnail_url:
          type: string
          format: uri
    Event:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
          format: date
        venue:
          type: string
        description:
          type: string
        source_url:
          type: string
          format: uri
        language:
          type: string
...
    TranslateRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
        source_language:
          type: string
          enum: [hi, en]
        target_language:
          type: string
          enum: [hi, en]
      required:
        - text
        - source_language
        - target_language
    TranslateResponse:
      type: object
      properties:
        translated_text:
          type: string
        source_language:
          type: string
        target_language:
          type: string
        cached:
          type: boolean
    BatchTranslateRequest:
      type: object
      properties:
        texts:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
        source_language:
          type: string
          enum: [hi, en]
        target_language:
          type: string
          enum: [hi, en]
      required:
        - texts
        - source_language
        - target_language
    BatchTranslateResponse:
      type: object
      properties:
        translations:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              cached:
                type: boolean
        source_language:
          type: string
        target_language:
          type: string
...
    SearchAnalytics:
      type: object
      properties:
        total_searches:
          type: integer
        unique_users:
          type: integer
        average_result_count:
          type: number
        average_response_time_ms:
          type: number
        top_search_terms:
          type: array
          items:
            type: object
            properties:
              term:
                type: string
              count:
                type: integer
        results_by_content_type:
          type: object
          additionalProperties:
            type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - INVALID_LANGUAGE
                - INVALID_AUDIO_FORMAT
                - UNAUTHORIZED
                - TOKEN_EXPIRED
                - FORBIDDEN
                - API_KEY_REVOKED
                - NOT_FOUND
                - DUPLICATE
                - PAYLOAD_TOO_LARGE
                - PROCESSING_FAILED
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
                - UPSTREAM_ERROR
                - SERVICE_UNAVAILABLE
                - MODEL_LOADING
                - UPSTREAM_TIMEOUT
            message:
              type: string
            details:
              type: object
            request_id:
              type: string
              format: uuid
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
...